======*****PHƯƠNG TIỆN*****===== 1.  Thống kê các booking theo trạng thái order     + Biểu đồ bánh     + Lọc theo năm     + Thể hiện bằng %
2.  Thống kê tổng số lượng khách hàng của chuyến xe
    + Biểu đồ cột ngang
    + Lọc theo năm
    + Thể hiện bằng số lượng

3.  Doanh thu hãng xe
    + Biểu đồ cột dọc
    + Lọc theo năm
    + Show hết 12 tháng của 1 năm
    + So sánh năm được chọn với năm trước

+++ src/main/java/com/main/traveltour/dto/agent/transport/StatiscalTransportBrandDto.java
@@ -13,5 +13,5 @@ public class StatiscalTransportBrandDto {
     private Integer month;
     private String formLocation;
     private String toLocation;
-    private Long maxAmount;
+    private BigDecimal maxAmount;
 }
